be/pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Productivity + Education App Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "google-generativeai>=0.8.3",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.32.1",
    "google-genai>=1.36.0",
]
